Move response interfaces out of implementation files to prevent leaking the implementation into type docs.
