+++ promise.d.ts
@@ -13,38 +13,27 @@ declare namespace simplegit {
 
    type SimpleGit = simpleGit.SimpleGit;
 
-   type Options = types.Options;
-
-   type LogOptions<T = types.DefaultLogFields> = types.LogOptions<T>;
-
    // errors
    type GitError = errors.GitError;
    type GitResponseError<T> = errors.GitResponseError<T>;
    type TaskConfigurationError = errors.TaskConfigurationError;
 
    // responses
-   // ---------------------
-   type BranchSummary = types.BranchSummary
-
-   type CleanSummary = types.CleanSummary;
-
+   type BranchSummary = resp.BranchSummary
+   type CleanSummary = resp.CleanSummary;
    type CleanMode = types.CleanMode;
-
-   type CommitSummary = resp.CommitSummary;
-
-   type MergeSummary = resp.MergeSummary;
-
-   type PullResult = resp.PullResult;
-
-   type FetchResult = resp.FetchResult;
-
-   type StatusResult = resp.StatusResult;
-
    type DiffResult = resp.DiffResult;
-
+   type FetchResult = resp.FetchResult;
+   type CommitSummary = resp.CommitSummary;
+   type MergeSummary = resp.MergeSummary;
+   type PullResult = resp.PullResult;
+   type StatusResult = resp.StatusResult;
    type TagResult = resp.TagResult;
 
+   // types
    type outputHandler = types.outputHandler
+   type LogOptions<T = types.DefaultLogFields> = types.LogOptions<T>;
+   type Options = types.Options;
 
 }
