+++ typings/index.d.ts
@@ -1,8 +1,8 @@
 import { SimpleGit } from './simple-git';
 
+export { SimpleGit };
 export * from './errors';
 export * from './response';
-export * from './simple-git';
 export * from './types';
 
 export declare function gitP(basePath?: string): SimpleGit;
