+++ typings/types.d.ts
@@ -1,8 +1,4 @@
 
-export { BranchDeletionBatchSummary, BranchDeletionSummary } from '../src/lib/responses/BranchDeleteSummary';
-export { BranchSummary, BranchSummaryBranch } from '../src/lib/responses/BranchSummary';
-export { CleanSummary } from '../src/lib/responses/CleanSummary';
-export { ConfigListSummary, ConfigValues } from '../src/lib/responses/ConfigList';
 export { RemoteWithoutRefs, RemoteWithRefs } from '../src/lib/responses/GetRemoteSummary';
 export { LogOptions, DefaultLogFields } from '../src/lib/tasks/log';
